Testing: Move installation functions into `agents` package
